Tidy up some logging, fix reported BIC discrepancy

+++ integration_tests/cleanup_files.sh
@@ -1,3 +1,12 @@
 # remove subdirectories
-find . -mindepth 2 -maxdepth 2 -type d -exec rm -r '{}' \;
-find . -mindepth 2 ! -name '*.pyc' ! -name '*.py' ! -name '*.txt' -exec rm '{}' \;
+DIRNAME=${PWD##*/}
+EXPECTDIRNAME='integration_tests'
+if [ $DIRNAME = $EXPECTDIRNAME ]
+then
+  echo 'Cleaning up'
+  find . -mindepth 2 -maxdepth 2 -type d -exec rm -r '{}' \;
+  find . -mindepth 2 ! -name '*.pyc' ! -name '*.py' ! -name '*.txt' -exec rm '{}' \;
+else
+  echo 'Careful! Calling an extensive "rm" in unknown directory!!'
+  echo 'Nothing happened... this time.'
+fi
